pyproject.toml
[project]
name = "thinkelearn"
version = "0.1.0"
description = "THINK eLearn"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.0",
    "django-storages[s3]>=1.14.6",
    "gunicorn>=23.0.0",
    "mailtrap>=2.4.0",
    "psycopg[binary]>=3.2.9",
    "sentry-sdk[django]>=2.31.0",
    "stripe>=12.0.0",
    "twilio>=9.6.3",
    "wagtail>=7.0.1",
    "wagtail-lms>=0.4.0",
    "whitenoise>=6.9.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.8.4",
    "mypy>=1.14.1",
    "django-stubs>=5.2.1",
    "types-requests>=2.32.0.20241211",
]
test = [
    "coverage>=7.9.1",
    "factory-boy>=3.3.3",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
    "wagtail-factories>=4.2.1",
]
security = [
    "safety>=3.3.1",
    "bandit[toml]>=1.8.0",
]

[tool.ruff]
target-version = "py313"
line-length = 88
extend-exclude = [
    "migrations",
    "venv",
    ".venv",
    "node_modules",
    "static",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"settings/*.py" = ["F403", "F405"]
"*/migrations/*.py" = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]
disallow_untyped_defs = false
disallow_any_generics = false
strict_optional = false

[[tool.mypy.overrides]]
module = [
    "*.models",
    "*.admin",
    "*.migrations.*",
    "*.tests.*",
    "*.settings.*",
]
ignore_errors = true

[tool.django-stubs]
django_settings_module = "thinkelearn.settings.dev"

[tool.bandit]
exclude_dirs = ["tests", "migrations", "venv", ".venv", "node_modules"]
skips = ["B101", "B601"]

[tool.coverage.run]
source = ["."]
omit = [
    "*/migrations/*",
    "*/venv/*",
    "*/.venv/*",
    "*/node_modules/*",
    "manage.py",
    "*/settings/*",
    "*/tests/*",
    "conftest.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "thinkelearn.settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*", "*Tests"]
python_functions = ["test_*"]
testpaths = ["home/tests", "blog/tests", "communications/tests", "portfolio", "lms"]
addopts = [
    "--verbose",
    "--tb=short",
    "--strict-markers",
    "--strict-config",
    "--cov=home",
    "--cov=blog",
    "--cov=communications",
    "--cov=portfolio",
    "--cov=lms",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=50",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::django.utils.deprecation.RemovedInDjango60Warning",
    "ignore::PendingDeprecationWarning",
]

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
include = [
    "home*",
    "blog*",
    "communications*",
    "search*",
    "thinkelearn*",
    "lms*",
    "payments*",
]
